[fixed] instance where int overflow was causing stationary healer agents to wait indefinitely

+++ MMOCoreORB/src/server/zone/objects/creature/ai/LuaAiAgent.cpp
@@ -349,7 +349,7 @@ int LuaAiAgent::stopWaiting(lua_State* L) {
 }
 
 int LuaAiAgent::setWait(lua_State* L) {
-	int wait = (int) lua_tonumber(L, -1);
+	uint64 wait = (uint64) lua_tonumber(L, -1);
 
 	Locker locker(realObject);
